Add player to list on join if the have the right permission.
